Home/requirements.txt
aiosmtplib==3.0.1
annotated-types==0.6.0 
anyio==4.2.0       
bcrypt==3.2.0       
certifi==2023.11.17  
cffi==1.16.0      
click==8.1.7       
colorama==0.4.6       
confluent-kafka==2.3.0       
cryptography==42.0.1      
dnspython==2.4.2       
ecdsa==0.18.0      
email-validator==2.1.0.post1 
exceptiongroup==1.2.0
fastapi==0.108.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.12
passlib==1.7.4
pip==23.3.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
secure-smtplib==0.1.1
setuptools==58.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0